requirements.txt
pyautogui
opencv-python
pytesseract
numpy
pandas
python-dotenv